symbol_resolve: use the static compilation.
Use the static compilation for symbol_resolve tools. The old version of
libelf have BUGs which result in the tools can not deal with relocated
functions correctly.

We can't ensure that all of the oneline systems have the high version
libelf. So using the static compilation with the high version of libelf.

Signed-off-by: Erwei Deng <erwei@linux.alibaba.com>
Reviewed-by: Yihao Wu <wuyihao@linux.alibaba.com>