src/x86: Fixed dynamic springboard when stack size is mismatched
stack size of __schedule is likely to be different in vmlinux and plugsched.
This happens when they're built with different version of gcc, or __schedule
is modified in plugsched.

Since the new __schedule() share code with bottom half of old __schedule,
stack mismatch will lead to kernel panic. So, we align the stack size to
solve it.

Two important skill:

1. Using no-omit-frame-pointer attribute to store frame pointer to %RBP;
2. To keep the clobbered register constant, tell gcc to clobber all callee
   saved register. Both vmlinux and plugsched need it.
   Fortunately, the __schedule function (in vmlinux) is complex enough, so
   all callee saved register is clobbered.

We use a script to analyzes vmlinux to record its stack layout. And we patch
__schedule(plugsched) to do the fixup according to what we recorded.

Co-developed-by: Erwei Deng <erwei@linux.alibaba.com>
Co-developed-by: Shanpei Chen <shanpeic@linux.alibaba.com>
Reviewed-by: Yihao Wu <wuyihao@linux.alibaba.com>