src: Simplify the sched rebuild.
Simplify the sched rebuild process. It is a better way to use the
de/enqueue function to clear and rebuild scheduler instead of by
manual. The de/enqueue can do the data synchronization work
automatically and thoroughly. Previously, we focused only on the CFS
scheduler, but now we can rebuild the other schedulers.

The de/enqueue function can help us rebuild the scheduler very well, but
they can't deal with the case that a cfs_rq is throttled. So we need to
solve this case ourselves.

Signed-off-by: Erwei Deng <erwei@linux.alibaba.com>
Reviewed-by: Yihao Wu <wuyihao@linux.alibaba.com>
Reviewed-by: Shanpei Chen <shanpeic@linux.alibaba.com>