src: Further optimize the rebuild use the set_rq_on/offline
The set_rq_offline can help us to clean the bandwidth control of cfs or
rt automatically and correctly. So we use this function to do some work
instead of by ourselves.

In the rebuild process, hrtimers of bandwidth control should be
restarted, otherwise, all of the group using BWC are throttled and not
scheduled.

Signed-off-by: Erwei Deng <erwei@linux.alibaba.com>
Co-Developed-by: Shanpei Chen <shanpeic@linux.alibaba.com>
Reviewed-by: Shanpei Chen <shanpeic@linux.alibaba.com>