border: Cleanup. Use fn_symbol_classify as set only.
It's kind of abusing dynamic type right now. Sometimes fn_symbol_classify
is list, sometimes it's converted to set. This is very confusing.

This patch sorts out the fact, that fn_symbol_classify is in essence a set,
not a list. This patch removes the usage where it is used as a list.

Signed-off-by: Yihao Wu <wuyihao@linux.alibaba.com>
Reviewed-by: Shanpei Chen <shanpeic@linux.alibaba.com>