src: Remove the retry process in load/unload routines.
The retry processes should be handled in user mode rather than kernel
mode. So I remove the retry process from load/unload routines. But there
is a BUG which is that, if stack check failed, the system would panic
because the module_exit function can not return errors. I will fix the
BUG in next patch.

Signed-off-by: Erwei Deng <erwei@linux.alibaba.com>
Reviewed-by: Yihao Wu <wuyihao@linux.alibaba.com>