src: Use procfs interface to control the load/unload routines.
The procfs handler can return a error, so I use the procfs to control
the module's load/unload routine. If there are anything failed, the
procfs handler can return errors to the user, not like the module_exit,
which can break off the routine and retry the operation.

The procfs interface is /proc/sys/kernel/scheduler_enable. Usage is
below:

Enable scheduler:  echo 1 > /proc/sys/kernel/scheduler_enable
Disable scheduler: echo 0 > /proc/sys/kernel/scheduler_enable

And the default value is "1" when install the rpm-package. If someone
want to remove the rpm or the scheduler module, must disable the module
firstly, and then can remove the rpm or module.

This patch dose not change the process of rpm-package install/remove, I
will fix it in next patch.

Signed-off-by: Erwei Deng <erwei@linux.alibaba.com>
Reviewed-by: Yihao Wu <wuyihao@linux.alibaba.com>