tools: Build symbol_resolve together with scheduler in scheduler tree
symbol_resolve was a stand-alone tool. So it was built from plugsched tree.
But it reads sched_outsider.h as an input file at compile time now, so we
must build it from scheduler tree now.

Of course we can make symbol_resolve read this file at runtime rather than
compile time. But considering it runs on production systems, compile time
is preferred for performance concerns.

This change makes symbol_resolve tool not stand-alone anymore. It should be
be placed where plugsched init works (in the scheduler tree)

Signed-off-by: Yihao Wu <wuyihao@linux.alibaba.com>
Reviewed-by: Shanpei Chen <shanpeic@linux.alibaba.com>