src: Use symbol_resolve instead of kallsyms_lookup_name to find function
I use the symbol_resolve tool to find the function address defined in
vmlinux instead of kallsyms_lookup_name because the symbol_resolve tools
use the symbol position to deal with the undefined funcions. This
methods can also save the time of loading module initialization.

I put the interface functions and function-pointers to the
sched_outsider.h file, and rename the file to undefined_functions.h,
which can indicates the purpose of this file.

Signed-off-by: Erwei Deng <erwei@linux.alibaba.com>
Reviewed-by: Yihao Wu <wuyihao@linux.alibaba.com>