border: Support gcc-9 for both init and build
We have supported gcc-6.5.1 previously. This patch supports gcc-9.

It fixes one and only one impatibility issue for gcc-9. That is, attribute
names are not canonicalized before gcc-8.1.0, which is undesired. And it's
fixed by gcc upstream in 8.1.0.

NOTE we support both canonicalized names and legacy names in this patch.

See this link for more information.
https://gcc.gnu.org/legacy-ml/gcc-patches/2017-06/msg02173.html

Signed-off-by: Yihao Wu <wuyihao@linux.alibaba.com>
Reviewed-by: Shanpei Chen <shanpeic@linux.alibaba.com>