src: Remove kbuild.patch
plugsched aims to cover a wide range of kernels (from 3.10 ~ 5.10), patch
files makes it hard to be portable. So better remove some of them.

BTW remove lazy compiling speedup. Because it breaks a lot of things, which
requires dirty fixing. Especially when portting to different kernels. Life
is easier after removing it.

Signed-off-by: Yihao Wu <wuyihao@linux.alibaba.com>
Reviewed-by: Shanpei Chen <shanpeic@linux.alibaba.com>