contrib: add Dockerfile for aarch64
Although docker has supported buildx, which can build image for different
architectures with a single Dockerfile, but older versions of docker is
still widely used. So we duplicate this file to make it a general solution
to build docker image for aarch64.

It's based on Alinux3 rather than Anolis8, because the kernel should be
compiled with gcc>=9. Otherwise CONFIG_HAVE_STACKPROTECTOR is turned off
automatically. And Anolis8 doesn't carry gcc as much new.

Signed-off-by: Yihao Wu <wuyihao@linux.alibaba.com>