config: springboard and stack balancing for aarch64
Springboard on aarch64 is similar to that on x86_64. But stack balancing
is a little more complex. Because fp and retaddr is stored on the top
of the stack for aarch64, but on the bottom of the stack for x86_64.

So if we balance the stack, we should move the data stored near [sp] too
along with sp itself. Otherwise data stored near [sp] will be flushed
unexpectedly.

Note that aarch64 has 10 callee-saved registers, x19-x28 and fp. But we
can't add fp to the clobber.

Co-developed-by: Shanpei Chen <shanpeic@linux.alibaba.com>
Signed-off-by: Yihao Wu <wuyihao@linux.alibaba.com>