border: fix gcc mangling
GCC mangles symbol names for various purposes. For example, it adds
.isra.xxx suffix to function names, if there are some unused parameters,
which can be eliminated/optimized. Other suffixes can be .part.xxx,
.cold.xxx, .constprop.xxx, .[numbers], .constprop.

This patch clarifies how does this behaviour influence sympos calculation
and symbol resolve.

GCC performs optimization under the context of kernel building, without
knowing existence of out-of-tree modules. So the most important constraint
is, modules shouldn't call those optimized functions. Otherwise the system
may become unstable or even crash. For plugsched, this constraint
corresponds to symbol resolve.

So how to tell which functions in vmlinux aren't optimized? Just check if it
ends with these suffixes. If they don't end with any GCC mangling suffix,
we can be sure they're not optimized, and safe to do symbol resolve.

Another problem is, if some functions were optimized, would the original
unoptimized symbol be left over? The answer is, /.part/.cold/ do but
/.isra/.constprop/ don't.

So, if some functions are to be symbol resolved, and it's already optimized
by Constant Propagation or IPA SRA, then it will eventually fail to be
loaded into the kernel. This might be fixed in the future, if it becomes a
severe problem in the future. For now we leave it be.

Apart from symbol resolve, there's another constraint. What sympos should be
assigned to the tainted_functions? Tainted_functions will be used to check
confliction with kpatch in production environment later. Contrary to
symbol_resolve, which replace scheduler symbols reference with vmlinux
symbols, tainted_functions do exactly the reverse, it replace vmlinux
symbols with scheduler symbols.

Tainted_functions must replace those non-optimized symbols in vmlinux. For
example, if theres a func1 and a func1.part.1 in vmlinux, and only a func1
in the scheduler.ko. We can be sure scheduler.ko replaced func1 in vmlinux
instead of func1.part.1 in vmlinux. Because func1 in both scheduler.ko and
vmlinux aren't optimized while func1.part.1 is optimized.

At last, .[numbers] is the suffix of static local variables. It's not very
urgent bug because the scheduler doesn't have many static local variables
yet.

Conclude from above, GCC mangled names is not of our concern. So we can
ignore them both when assigning symposes and symbol resolve. And this
patch fixes .part., .cold sufixes, and avoids hidden bugs of .isra and
.constprop, and leaves .[numbers] bugs as they be.

Reported-by: Erwei Deng <erwei@linux.alibaba.com>
Signed-off-by: Yihao Wu <wuyihao@linux.alibaba.com>