contrib: fix the hotfix_conflict_check.sh script
Conflict checking aims to detect the confliction between **newly**
tainted functions and existing tainted functions.

Confliction among existing tainted functions is not of our concern. This
patch excludes this type of confliction.

For example, if kpatch patched the same function twice, the conflict
detection will fail when installing the scheduler rpm, even if the
tainted functions of scheduler rpm do not conflict with the existing
tainted functions.

Signed-off-by: Erwei Deng <erwei@linux.alibaba.com>