contrib: fix the bug when remove the scheduler rpm
The system maybe crashed before the installation files of the RPM
could be written to disk if the scheduler.ko module has BUGs. And after
reboot, removing the RPM will fail because the scheduler-install script
file can not be found. See the error log:

  # rpm -e scheduler-xxx
  Uninstalling scheduler
  /var/tmp/rpm-tmp.HgRsuX: line 4: /usr/local/bin/scheduler-install: No
  such file or directory
  error: %preun(scheduler-xxxxxxx.x86_64)scriptlet failed, exit status 1
  error: scheduler-xxxxxxx.x86_64: erase failed

So, it is necessary to write the files to disk before installing the
kernel module.

Signed-off-by: Erwei Deng <erwei@linux.alibaba.com>