src: use a more generic API to get symbols' size
I replace the get_ksymbol with kallsyms_lookup_size_offset for the
following two reasons:

1) kallsyms_lookup_size_offset is more generic than get_ksymbol. The
   get_ksymbol no longer exists in kernel v5.10.

2) Using get_ksymbol requires RCU lock protection, otherwise there will
   be a warning.

After test, the time cost of kallsyms_lookup_size_offset is more
than get_ksymbol, but only a few hundred microseconds. And getting
symbols' size is not in stop machine context. So, they are accepted.

Reported-by: Qiao Ma <mqaio@linux.alibaba.com>
Signed-off-by: Erwei Deng <erwei@linux.alibaba.com>
Co-developed-by: Qiao Ma <mqaio@linux.alibaba.com>