tools: search springboard for aarch64
We search in the vmlinux assembly for x86_64

1. stack size of __schedule
2. the instruction next to "call __switch_to" which is called springboard

Besides them, we search for aarch64

3. the range of stack protector instruction sequence in __schedule

stackprotector_per_task is a lightweight canary checking capability of
aarch64. We'll disable the stackprotector_per_task for __schedule in the
following patches, and we'll explain why.

For now, we explain how to identify the instruction sequence for it, which
is "ldr ldr (any ){1,4}chk (ldp ){6}ret". It first loads two values from
memory, then compares these two values (there are multiple possible ways to
compare them, so we accept 1~4 arbitrary instructdions), if they differ
then chk (abbrevation for goto stack_chk_fail). Finally it does the
standard stack restoring and return.

This patch locates where this sequence locates.

Signed-off-by: Yihao Wu <wuyihao@linux.alibaba.com>