src: don't print "defined but not used function" warning message
For normal program, it's good choice to warn whenever a static
function is declared but not defined or a non-inline static
function is unused.

But for plugsched, interface & fn_ptr functions defined in scheduler.ko
is used for indiretion at runtime, rather than gcc compile stage. So,
it's better to disable Wunused-function.

Signed-off-by: Shanpei Chen <shanpeic@linux.alibaba.com>