src: support memory pool for percpu variables
There are two ways to implement percpu memory pools. One is to design a
percpu memory pool ourselves, and the other is to use the original
percpu mechanism of the kernel.

For the first way, we need to allocate a block of memory(such as a
array) in each numa node to store its own cpus' variables, and the base
address of allocated memory in each numa should be recorded in a array.
And then, the cpus on the nodes need to be renumbered because the cpu
numbers on the nodes are not consecutive. Finally, cpu accesses its own
variables through the following way:

    base_addr[cpu_to_node(cpu)][cpu_id_in_numa(cpu)];

There are two drawbacks to this way. First, We need to create a new set
of APIs for the percpu instead of using the kernel's APIs. Second,
accessing a percpu variable requires two memory-references.

For the second way, in the kernel, the maximum supported allocation size
is PCPU_MIN_UNIT_SIZE defined in include/linux/percpu.h that is 32KB in
anck-v4.19 for x86_64. I apply the maximum size as a large variable and
make that as a part of memory pool.

The address of each variables of CPUs is the base address plus a fixed
offset. Following this rule, we can assign multiple small variables
from a large variable. And if the large variable is not enough, apply
other large variables. The offset is stored in gs register(x86) and
tpidr register(arm), so, the cpu accessing its own variable just needs
one memory-reference.

In conclusion, I choose the second way.

Signed-off-by: Erwei Deng <erwei@linux.alibaba.com>