src: fix memory leak when use mempool
The memory pool is created in the function load_sched_routine which
is the callback of 'sysfs enable', but the destroy happens when
unloading the scheduler module. So there is a case which will result
in memory leak:

  # rpm -ivh scheduler-xxx-***********.yyy.x86_64.rpm  --> create pool A

  # echo 0 > /sys/kernel/plugsched/plugsched/enable    --> not destroy A

  # echo 1 > /sys/kernel/plugsched/plugsched/enable    --> create pool B

  # rpm -e scheduler-xxx                               --> destroy B

The mempool pool A is not destroyed (memory leak occurs).

This patch fixes this bug.

Signed-off-by: Xuchun Shang<xuchun.shang@linux.alibaba.com>