docs: don't assume kernel version in case image updates
We previously assumes users have 4.19.91-25.2.an7.x86_64 kernel installed,
which is not the case. Because the default kernel for Anolis OS 7.9 ANCK is
constantly upgrading. If we still make this assumption, users may encounter
the problem that downloaded src.rpm is older than the kernel which users
are using.

Also, we even have both 4.19.91-25.1.al7.x86_64 and 4.19.91-25.2.al7.x86_64
written in the README. It definitely confuses users.

Signed-off-by: Yihao Wu <wuyihao@linux.alibaba.com>