tests: add cpu_throttle test
CFS bandwidth control is one of the representative features of CFS core.
This test ensures the inheritance of CFS bandwidth control related data
(consistency is guaranteed before and after module loading/unloading).

The test steps are as follows:

1) Create a cpu cgroup, run busyloop (infinite loop) in it,
set cpu.period and cpu.quota , check if the CPU% is correct.

2) Load the module and check the correctness of the CPU%
of the busyloop process.

3) Modify the value of quota and check the correctness of CPU%.

4) Unload the module and check the correctness of CPU%.

Co-developed-by: Erwei Deng <erwei@linux.alibaba.com>
Signed-off-by:  Xuchun Shang <xuchun.shang@linux.alibaba.com>