border: exclude __init functions from tainted_functions
They are not tainted at all, because they are removed from the memory long
time ago. We sure can't taint them at all. This patch simplifies the list
of tainted_functions.

init is removed from optimized_out, because whether optimized or not, we
don't need their definitions anymore.

BTW, simplify (!in('interface') && !in('fn_ptr')) to !in('border').

Signed-off-by: Yihao Wu <wuyihao@linux.alibaba.com>