config: add wake_up_idle_ht as an interface function
Due to springboard technique, __schedule is divided into two halves. The
top half can be seen as an interface function. And the bottom half can be
seen as an outsider function.

wake_up_idle_ht is called by the outsider half of __schedule. So it
should've been infected as an outsider too. But the current sched_boundary
algorithm sees __schedule as a whole interface function. So it was wrongly
classified as an insider. However the modification won't take effect.
What's worse, there maybe two versions of wake_up_idle_ht running in the
system.

To avoid confusion, and also allow users to modify this it, we add it as
an interface function.

Signed-off-by: Yihao Wu <wuyihao@linux.alibaba.com>