src: Migrating from python2 to python3
Now the python version of plugsched is Python 2.x which is no
longer supported. So migrating to Python 3.x is necessary.

The differences of features used in plugsched between Python 2.x
and Python 3.x are as follows:

    1) map() returns an iterator, not a list in Python 3.x, So
      we need to use list to wrap its return values.

    2) filter() is the same as map() in Python 3.x.

    3) unicode is replaced with str in Python 3.x.

    4) iteritems is replaced with items in Python 3.x.

    5) sublist of function parameter are no longer supported
      in Python 3.x.

And there is a python-related issue described in
https://fedoraproject.org/wiki/Changes/Making_sudo_pip_safe
The key point releated with plugsched is that the /usr/local path
will not be included in python sys.path when dealing with rpmbuild.
So we need to add the PYTHONPATH manually before rpmbuild start
which implemented in add_python_path(cli.py).

Signed-off-by: Xuchun Shang <xuchun.shang@linux.alibaba.com>