cli: put off copying the vmlinux and Module.symver files
Commit 2ed82330(src: fix the bug kernel modules don't inflect outsiders)
can generate the new vmlinux and Modules.symvers files after the collect
stage in the init process, which can rewrite these two files copied from
kernel-devel and kernel-debuginfo and we can not get the right symbols'
informations of the vmlinux.

I try to modify the Makefile to find a way that can compile kernel and
modules without generating the vmlinux and Modules.symvers, but failed.

So, copying the vmlinux and Modules.symvers files from kernel-devel and
kernel-debuginfo after the collect stage is a simple way to fix the
issue above, and then the two files can cover the generated files.

Fixes: 2ed82330 (src: fix the bug kernel modules don't inflect outsiders)
Signed-off-by: Erwei Deng <erwei@linux.alibaba.com>