src: remove the *.stub.o intermediate procedure
This is done to support the kernel-5.10.

The Makefile of kernel provides the objcopy command to deal with the *.o
files, so we need not use the *.stub.o intermediate procedure. In
kernel-5.10, *.cmd files cannot be generated when compiling the
scheduler module. See the error log:
  ...
  ...
  BJCOPY /tmp/work/scheduler-test/kernel/sched/mod//core.stub.o
  OBJCOPY /tmp/work/scheduler-test/kernel/sched/mod//fair.stub.o
  LD [M]  /tmp/work/scheduler-test/kernel/sched/mod//scheduler.o
  MODPOST /tmp/work/scheduler-test/kernel/sched/mod//Module.symvers
  /tmp/work/scheduler-test/kernel/sched/mod//.core.stub.o.cmd: No such file or directory
  ...
  ...

This patch can fix the error above.

Co-developed-by: Yihao Wu <wuyihao@linux.alibaba.com>
Signed-off-by: Erwei Deng <erwei@linux.alibaba.com>