border: move fake.c to kernel/sched
This is done to support the kernel-5.10.

In kernel 4.19, fake.c can be compiled using the "make fake.o"
command(the fake.c is copied to the root dir of kernel), but it can not
be compiled in kernel 5.10.  The kbuild does not seem to allow
compiling a source file arbitrarily that have not been added to a
Makefile. See the error log:

  ...
  ...
    CC      kernel/sched/idle.o
    CC      kernel/sched/pelt.o
    CC      kernel/sched/rt.o
    CC      kernel/sched/stats.o
    CC      kernel/sched/stop_task.o
    CC      kernel/sched/topology.o
  make[4]: *** No rule to make target 'kernel/sched/fake.o'.  Stop.
  make[3]: *** [scripts/Makefile.build:472: __build] Error 2
  make[2]: *** [scripts/Makefile.build:497: kernel/sched] Error 2
  ...
  ...

To compiling fake.c, I move the fake.c to kernel/sched directory and add
fake.o target to the Makefile by pre_extract.patch in kernel 5.10
config.

Signed-off-by: Erwei Deng <erwei@linux.alibaba.com>