border: add __used attribute to function pointers
Some function pointers are not used anywhere and may be optimized and
deleted by gcc, but they will replace corresponding functions in
kernel. So I add the __used attribute to function pointers to aviod
this case. See the error log in kernel 5.10:

  ...
  ...
  scheduler: Unknown symbol __mod_set_next_task_rt (err -2)
  scheduler: Unknown symbol __mod_rt_task_fits_capacity (err -2)
  ...
  ...

The root cause of this problem is that set_next_task_rt callback is
a member of sched_class and sched_class_* are shared variables(just
declared but not defined), so the set_next_task_rt callback is
optimized.

If some function pointers are used only in module, they can be inside
functions which really solves this problem. Now this is just a temporary
solution that won't last long.

Signed-off-by: Erwei Deng <erwei@linux.alibaba.com>