border: move optimized function pointers outside
Some function pointers are optimized as inline functions, so their
symbols can not be found in vmlinux. Now the gcc python plugin does
not classify function pointers. So I move the optimized function
pointers outside conservatively and temporarily.

In kernel 5.10, see the error log:
  ...
  ...
  scheduler: Unknown symbol __pushable_less (err -2)
  ...
  scheduler: Unknown symbol __entity_less (err -2)
  ...
  scheduler: Unknown symbol __dl_less (err -2)
  ...
  ...

In fact, some function pointers are used only in module, so they should
be inside functions. Now this is just a temporary solution that won't
last long.

Signed-off-by: Erwei Deng <erwei@linux.alibaba.com>