border: clean up the extraction of EXPORT_*SYMBOL
This is done to support the kernel 5.10.

In kernel 5.10, EXPORT_TRACEPOINT_SYMBOL macro can be expanded to eight
declarations of variables and we don't need them in fact because the
tracepoint are shared data. See the following extracted of the
EXPORT_TRACEPOINT_SYMBOL(pelt_cfs_tp):

  ...
  ...
  extern const char __kstrtab___tracepoint_pelt_cfs_tp[];
  extern const char __kstrtabns___tracepoint_pelt_cfs_tp[];
  extern const char __kstrtab___traceiter_pelt_cfs_tp[];
  extern const char __kstrtabns___traceiter_pelt_cfs_tp[];
  extern const char __kstrtab___SCK__tp_func_pelt_cfs_tp[];
  extern const char __kstrtabns___SCK__tp_func_pelt_cfs_tp[];
  extern const char __kstrtab___SCT__tp_func_pelt_cfs_tp[];
  extern const char __kstrtabns___SCT__tp_func_pelt_cfs_tp[];
  ...
  ...

EXPORT_TRACEPOINT_SYMBOL macro references the EXPORT_SYMBOL marco, and
for the EXPORT_SYMBOL macro, it declares some variables and defines two
static variables explicitly in kernel 4.19. But when the kernel version
>= 5.6.0, these two static variables are defined by assembly in asm
macro and recognized as global symbol, so they are extracted into
core.c.

So the EXPORT_SYMBOL macro can be expanded to two declarations in kernel
5.10:

  extern const char __kstrtab_******;
  extern const char __kstrtabns_******;

This may be a gcc bug, and it is not necessary to extract them out.

Signed-off-by: Erwei Deng <erwei@linux.alibaba.com>