src: rename bsearch to func_addr_bsearch
The function name "bsearch" in helper.h would conflict with the function
which has the same name in lib/bsearch.c. When the compiler check is
weak, helper.h would replace lib/bsearch.c all the time, which is not
expected for kernel source code.

Rename the function in helper.h to func_addr_bsearch to avoid conflicts.

Signed-off-by: Tianchen Ding <dtcccc@linux.alibaba.com>