src: deleting information collection for kernel modules
Since kernel modules can only reference exported symbols in vmlinux,
we only need to take care of exported symbols. According to exported
functions rule, a exported function can't be an internal function. So
there is no need to collect information for kernel modules.

BTW, with this patch, the init stage can be speed up a lot.

Signed-off-by: Shanpei Chen <shanpeic@linux.alibaba.com>