src: get stack size of __schedule in module for stack balance
Now, the stack balance technique relies on the stack frame register.
But stack frame register of __schedule function may be optimized for
example in kernel 5.10. So I use other way to do this work that is
compiling core.c file twice and getting the stack size in module.

The way of stack balancing becomes following:

        add sp stack_size_of_module
        sub sp stack_size_of_vmlinux

This patch does two things, and the one is compiling the core.c twice,
and the other is using the springboard_search.sh script to get the
stack size of __schedule in core.stub.o generated by the first
compilation and return the value to STACKSIZE_MOD macro.

It's important to note that the layout of callee saved registers on
stack of __schedule function in module must be same as it in vmlinux.

Signed-off-by: Erwei Deng <erwei@linux.alibaba.com>