contrib: automatically tag when updating version file
The version file was forgotten to update for quite a while, because it's
really a burden to manage both the tag and the version file at the same
time. So it's good if we only need manage one of them.

Given the fact tag comes after commit. So make a commit trigger a tag is
more practical.

Signed-off-by: Yihao Wu <wuyihao@linux.alibaba.com>