tools: fix some compilation errors for kernel 4.19 aarch64
There are some compilation errors when building the scheduler module
for arm64 kernel-4.19. See the first error log:

  /root/scheduler/kernel/sched/mod//main.c: In function
  'disable_stack_protector':
  /root/scheduler/kernel/sched/mod//main.c:137:35: error:
  'STACK_PROTECTOR' undeclared (first use in this function)
    137 |  void *addr = __orig___schedule + STACK_PROTECTOR;
        |                                   ^~~~~~~~~~~~~~~

  /root/scheduler/kernel/sched/mod//main.c:137:35: note: each
  undeclared identifier is reported only once for each function it
  appears in /root/scheduler/kernel/sched/mod//main.c:139:14: error:
  'STACK_PROTECTOR_LEN' undeclared (first use in this function)
    139 |  for (i=0; i<STACK_PROTECTOR_LEN; i++, addr+=4)
        |              ^~~~~~~~~~~~~~~~~~

For the CONFIG_STACKPROTECTOR_PER_TASK, it will be added to config files
automatically when gcc version >= 9.1.0. The gcc version of
plugsched-sdk is 10.2.1 but 8.5.1 to build the kernel of anolis 8.6.
So, the reason of this error is that the springboard_search.sh tool uses
the /usr/src/kernels/release-kernel/.config to determine whether the
CONFIG_STACKPROTECTOR_PER_TASK is set, but this config is added to
config file later when building the kernel in init stage, which
causes that the value of this config is "N" in the tool and "Y" in
main.c.

So the main.c should not use the CONFIG_STACKPROTECTOR_PER_TASK. We can
reuse the STACK_PROTECTOR macro generated by springboard_search tool in
main.c.

I found another bug when fixing the bug above. See the error log:

  /root/scheduler/kernel/sched/mod//main.c:137:50: error: expected
  expression before ';' token
    137 |  void *addr = __orig___schedule + STACK_PROTECTOR;
        |                                                  ^

The reason of this error is that, the STACK_PROTECTOR and
STACK_PROTECTOR_LEN have the invalid value. See the following in
Makefile:

  ccflags-y += -DSTACK_PROTECTOR=
  ccflags-y += -DSTACK_PROTECTOR_LEN=-7

The tool wants to match the "<__schedule+0x***>" in asm of __schedule
function. But it uses the last column incorrectly, because the last
column is a comment that is invalid.

Signed-off-by: Erwei Deng <erwei@linux.alibaba.com>