contrib: fix some bugs of dependent libraries
This patch fixes two bugs.

The first one is that, building the CLoader of pyyaml needs the
python3-Cython in Alinux3-aarch64, otherwise there is an error
of compilation and see the error log:

  gcc: error: yaml/_yaml.c: No such file or direc
  gcc: fatal error: no input files
  compilation terminated.
  error: command 'gcc' failed with exit status 1

For x86_64 of anolis 8.6, installing this package has no other effects.
But in order to reduce the differences in Dockerfiles, I install this
package in Dockerfile.x86_64 too.

The second one is that, the gcc-python-plugin needs use the 0.17-1.2
instead of 0.17-1.1 in Dockerfile.aarch64.

Signed-off-by: Erwei Deng <erwei@linux.alibaba.com>