src: fix the compilation errors for kernel-5.10
This patch fixes two compilation errors for kernel-5.10.

For the first one, using "make $(vmlinux-deps)" can cause the following
error:

  ...
  ar: fs/notify/built-in.a: No such file or directory
  make[1]: *** [scripts/Makefile.build:400: fs/built-in.a] Error 1
  ...

The vmlinux-deps variable includes some target files for example the
*/built-in.a. In kernel 5.10, if we build a single target, the target
of its subdir can not be built. For example, I build the fs/built-in.a
which will report this error:

  # make fs/built-in.a
  ...
  ...
  ar: fs/notify/built-in.a: No such file or directory
  make[1]: *** [scripts/Makefile.build:400: fs/built-in.a] Error 1

But if using the "make fs", all of the target objs will be built. So I
replace the vmlinux-deps target to vmlinux-dirs target, and after the
compilation, all of the target of vmlinux-deps can be built too.

So I think using vmlinux-dirs is a better way and the vmlinux-dirs is
very general and is still used in the latest kernel.

For the second one, see the error log when building the scheduler
module:

  ...
  ...
  make[3]: *** No rule to make target 'scripts/module.lds', needed by
  '/tmp/work/scheduler/kernel/sched/mod//scheduler.ko'. Stop.
  ...
  ...

Using the modules_prepare target can fix this bug.

After fix these two bugs above, I tested the init stage time for
kernel-4.19 in 64 CPUs system and the result looks good.

  raw time: 51.472s
  new time: 52.009s

Signed-off-by: Erwei Deng <erwei@linux.alibaba.com>