border: adapter code due to abandoning get_location()
For 'decl_end_line' attribute is abandoned, another solution should
be used to find the end line of variable definition.

For clang, the end of variable definition should be ";", this rule
can be used to search the end line. After adding this rule, merge_
down mechanism play a more general role in code extractin.

Signed-off-by: Shanpei Chen <shanpeic@linux.alibaba.com>