border: forbid redirecting mangled functions
interface, callback or sidecar functions are to be redirected when loading
the module. But if these functions are mangled by GCC, they lose the
original semantics in general cases, because mangling usually happens when
optimzing for some special cases. Redirecting them will lead to bugs.

Moreover, GCC usually leaves an unmangled symbol in the binary as well,
making it look safe to redirect this unmangled symbol. But is not safe.

So in the analyze phase, we check if interface, callback or sidecar
functions are mangled. If they do, the analyzing fails.

Signed-off-by: Yihao Wu <wuyihao@linux.alibaba.com>