tools: fix stackprotector search for aarch64
We search stackproctors by matching "<__schedule+OFFSET>" in the assembly,
where __schedule+OFFSET points to __stack_chk_fail.

11d1a5437a2c changes stack protector matching pattern from the last column
in the assembly to the 5th column. Because sometimes the jump instruction
used by stack protector is "b.ne", and b.ne has only one operator. But in
other cases, the instruction might be "cbnz", which has two operators.

To take both cases into account, we should check both the 5th and the 6th
column.

Fixes: 11d1a5437a2c ("tools: fix some compilation errors for kernel 4.19 aarch64")
Signed-off-by: Yihao Wu <wuyihao@linux.alibaba.com>