tools: use CONFIG_STACKPROTECTOR to indicate stack_chk_fail
As long as CONFIG_STACKPROTECTOR=y is present in the config file, we can be
sure that stack_chk_fail will be generated by GCC. And we must disable it
in this case. On the other hand, CONFIG_STACKPROTECTOR_PER_TASK=y has
nothing to do we stack_chk_fail at all.

We disccussed this previously, and came the conclusion above. However the
implementation was wrong in the end, which is strange.

Signed-off-by: Yihao Wu <wuyihao@linux.alibaba.com>