border: fix local symbol postion assert error
In many cases, a header file is included by more than one source
files. If a static function xxx defined in this header file are not
optimized during compiling, then there are several xxx symbols
in vmlinux.

According to 'Disagreement 2', all these xxx symbols are belongs
to the same header file. As a result, an Assert Fail is encountered.

For function xxx, any version (position) in vmlinux is equivalent,
so this patch fixes the assert error by simply delete the assert.

Fixes: 2257e93e ("border: fix gcc mangling")
Signed-off-by: Shanpei Chen <shanpeic@linux.alibaba.com>