border: make the function warning comment clear
The warning comment "IT'S NOT PART OF SCHEDMOD" is not clear for me.
If a function is not belong to new module, why don't move it out?

Actually, these functions are used by both vmlinux and scheduler
module, so they should be classified as outsider functions. Meanwhile,
their code is also optimized by GCC, and symbols can't be found in
vmlinux.

Signed-off-by: Shanpei Chen <shanpeic@linux.alibaba.com>