border: fix unnamed struct/enum strart location
First, let's recall what is main_variant, with an example,

    u64.main_variant -> long long unsigned int
    (const int).main_variant -> int
    (volatile int).main_variant -> int

`var_decl_start_loc` was to find the first line of the variable definition,
especially variables like `nohz`, which mixes type definition and variable
definition together. This special case can be identified by whether the
type has a name. Here's how it works,

    const struct {
        int x;
    } myvar1;

    myvar1.type -> const struct { int x; }
    myvar1.type.main_variant -> struct { int x; }       # it's a gcc type
    myvar1.type.main_variant.name -> None               # an anonymous type
    myvar1.type.main_variant.stub -> struct { int x; }  # it's a gcc decl
    myvar1.type.main_variant.stub.location -> 'struct {'
                                               ^
But the current method fails to handle this case,

    typedef struct { int x; } mytype;
    mytype myvar2;

    myvar2.type -> mytype
    myvar2.type.name -> mytype                   # not an anonymous type
    myvar2.type.main_variant -> struct { int x; }
    myvar2.type.main_variant.name -> None        # but mistook as anonymous

If the variable's type is not anonymous, we know the definition of the type
and the variable is separated. So we should use the variable's location
instead of the type's.

The way to fix this bug, is to check the name of the variable's type. And
the variable's main_variant's name is not concerned. After the bugfix, both
cases above can be handled correctly,

    const struct {
        int x;
    } myvar1;

    myvar1.type -> const struct { int x; }
    myvar1.type.name -> None                            # an anonymous type
    myvar1.type.main_variant -> struct { int x; }       # it's a gcc type
    myvar1.type.main_variant.stub -> struct { int x; }  # it's a gcc decl
    myvar1.type.main_variant.stub.location -> 'struct {'
                                               ^
    ---

    typedef struct { int x; } mytype;
    mytype myvar2;

    myvar2.type -> mytype (type)
    myvar2.type.name -> mytype        # not an anonymous type, stop here

Signed-off-by: Yihao Wu <wuyihao@linux.alibaba.com>