tools: stop supporting builtin springboard
Once we have some kernels customized for plugsched. And we patched these
kernels with builtin springboards. Now that dynamic springboard is mature
enough, and doesn't require any customizing to the kernel, we can delete
the support for builtin springboard.

Also, this patch fixes the bug, that plugsched fails to work on some of
these kernels. Because currently, dynamic_springboard.patch is composed
of stack-pivot and springboard. And for these kernels, stack-pivot is
not applied as well. Without stack-pivot, the scheduler is easily broken.

Signed-off-by: Yihao Wu <wuyihao@linux.alibaba.com>