tools: fix stack protector search
00a24b77eed7 ("tools: stop supporting builtin springboard") removes
"<__schedule>:" from asm_sequence, but still matches it when
searching stackprotector, which always fails.

First lets recall why we matches "<__schedule>:". This is a sample of
objdump -d

    /usr/lib/debug/lib/modules/4.19.91-26.an8.aarch64/vmlinux:   ...

    Disassembly of section .text:

    ffff0000089fb368 <__schedule>:
    ffff0000089fb368:	a9b77bfd 	stp	x29, x30, [sp, #-...
    ffff0000089fb36c:	d0004161 	adrp	x1, ffff000009229...

We convert disassembly to asm_sequence, so that we can easily grep
the desired pattern. To avoid objdump header information converted
to asm_sequence, we match "<__schedule>:" and set a "start" tag.
But since "<__schedule>:" is gone now as below,

    /usr/lib/debug/lib/modules/4.19.91-26.an8.aarch64/vmlinux:   ...
    Disassembly of section .text:
    ffff0000089fb368:	a9b77bfd 	stp	x29, x30, [sp, #-...
    ffff0000089fb36c:	d0004161 	adrp	x1, ffff000009229...

we should turn to match "Disassembly of section" now.

Fixes: 00a24b77eed7 ("tools: stop supporting builtin springboard")
Signed-off-by: Yihao Wu <wuyihao@linux.alibaba.com>