border: fix GCC optimization of fields accessing grouping
For `union bpf_attr`, when visiting attr->bpf_fd and attr->file_flags,
GCC optimizes these two u32 to one u64. i.e.

    attr->bpf_fd & 0xBEEF || attr->file_flags & 0xDEAD

becomes,

    tmp = *(struct unnamed_struct*)((unsigned long)attr+8)
    tmp | 0xdead0000beef

The current algorithm failed to handle this, because GCC thinks there are
only one ComponentRef, but it sees `bpf_fd` and `file_flags` as a whole
field, which is anonymous too. This leads to crash when sorting of fields
by their name.

This can be fixed by ignoring the field's name, and using a fixed fake
name `<unknown>` instead. In this case, the most important information is
union name bpf_attr, other names, such as bpf_fd and file_flags, can be
ignored if them are hard to get.

Signed-off-by: Yihao Wu <wuyihao@linux.alibaba.com>